Attempting to calculate toggling on timer loops (semi-working)
With CC value of 400 it kinda "works" (haven't checked the output, though),
but – at least in REPL – with CC of 100 the board becomes unreasonably
slow.

MAYBE it could work if not in REPL?

But a sane experiment to do from here could be to write a script to be run
as code.py, running e.g. 40 Hz tone with different CC values, and LISTEN.

UPDATE: Actually did a quick listen – one note sounds not bad. I mean, no
obvious artifacts or so, but didn't measure frequency.
Tried 40 Hz, 640 Hz, 2560 Hz. The smallest CC: 400.

NEXT: Maybe try implementing multiple tones and basic PWM (in that order).